Use tags when pushing to dockerImageDestination
Not pushing to a tag makes the image impossible to pull with a tag.

This is not really correct because we keep the Name and Tag in schema1
manifests unmodified, and older registry versions (in particular
OpenShift 1.1) do inspect those fields.  But, it is better than nothing,
and for OpenShift manifest uploads via the Docker API
( https://github.com/containers/image/pull/95 )a tag is required.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>