Add ImageDestination.Commit
This marks the image saving as "done".  This is a separate concept from
ImageDestination.Close, which is called whenever the destination
should no longer be used, even if the caller has given up on the upload
process.

The separation between Close and Commit alllows a natural, simple flow:
> dest := ref.NewDestination()
> defer dest.Close()
> ... // do anything, simply return on error
> dest.Commit()

whereas a single Close would be more difficult to use:
> dest := ref.NewDestination()
> succeeded := false
> defer func() {
>    dest.Close(succeeded)
> }()
> ... // do anything, simply return on error
> succeeded = true

Signed-off-by: Miloslav Trmač <mitr@redhat.com>