Add ManifestUpdateOptions.InformationOnly
Manifest conversion will need a fair amount of information about the
copied image and its layers.  Start by providing the ImageDestination
(to upload throwaway=empty_layer layers when converting to schema1),
and BlobInfos of the uploaded layers (in particular, providing the layer
size even if we didn't recompress, because schema1 does not store it and
schema2 needs it).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>