Don't enforce blobinfo in storage.ReapplyBlob
When reapplying a blob to an image that's being written to the storage
transport, don't enforce that the digest and size of the reapplied blob
match the blobinfo's values, if they were provided, because compression
makes the reapply method (where we reextract the layer, and if it was
originally compressed, recompress it) produce data with the same actual
content, but with a different compressed bitstream.

Like we were previously doing for digest values from the blobinfo, make
sure that we return the passed-in blob size if one was specified, else
return the value we computed.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>