Support testing dockerImageSource via github.com/dnaeon/go-vcr/recorder, add an example
This allows running the test against a live registry once, manually
inspecting the requests/responses to be as expected, and then running
all tests against a “recording” of the requests/responses, completely
off-line.

Also adds a simple test for GetManifest to demonstrate the use of the
mechanism.

The httpWrapper mechanism is intended purely for tests, and not intended to be
exposed to external users; they should use types.SystemContext.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>