Don’t limit atomic: push to schema1
Instead, just forward SupportedManifestMIMETypes to the underlying
docker: transport.  Now that we can do autodetection, it should be safe
to declare schema2 (or, in the future, OCI) support.

One notable difference is that SupportedManifestMIMETypes affects what
we configure the source to accept: e.g. when streaming from docker: to
s1-only atomic:, previously the conversion was done by the docker:
registry server, but now the source registry gives us schema2 and we do
the conversion, changing the manifest digest.  Right now that is a
difference, but soon(?) we will update the manifest to use the
destination’s name/tag, and then the manifest digest will change in any
case.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>