Disable the ostree transport on macOS
Set the containers_image_ostree_stub build tag automatically on macOS;
at least right now, ostree/src/libostree/ostree-linuxfsutil.c hard-codes
Linux-specific header files and ioctls, so we are fairly sure that it will
not be available no macOS.

Make the ostree subpackage also conditional on containers_image_ostree_stub ,
so that we don't even try to build it (and any reference to it fails with
"no buildable source files").

Finally, use $BUILDFLAGS also in (go list) when listing the subpackages,
so that we don't try to run tests in the ostree subpackage (and therefore
try to build it) when it has been disabled.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>