transports: make containers/storage stubbable
Currently containers/storage is only used in very specialised use-cases,
outside of the more generic usecase of converting to and from OCI
images. As a result, the fairly involved build dependencies are not
necessary for a large portion of users (for example, users that only use
skopeo as a tool for fetching OCI images and then use umoci or other
such tools to manipulate said images).

This copies the existing stubbing conventions for ostree.

Signed-off-by: Aleksa Sarai <asarai@suse.de>