storageImageDestination.computeID: move to manifest.ImageID()
Move most of the storageImageDestination.computeID() logic into an
ImageID() method of manifest types.

For Schema1, break it into a two-step process, so that we construct a
new configuration blob first, and digest it to produce the ID in a
wrapper.  Update image/manifestSchema1.convertToManifestSchema2() to use
that same configuration blob.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>