Change errors structure to handle them gracefully in docker
If commiting or reading Docker image fails the error message is verbose
and thus it's hard to handle the error gracefully.

This patch attempts to solve in the following way:
The `docker/distribution/registry/client.HandleErrorResponse` takes a
`http.Response` and returns a new error containing a human-readable
error message alongside with the status code.

In order to handle such errors, verify the error type and handle it:

if respErr, ok := errors.Cause(err).(errcode.Error); ok {
    // handle the error here.
}

This change is going to apply on docker images.

Signed-off-by: Boaz Shuster <ripcurld.github@gmail.com>