Move LayerInfosForCopy from UnparsedImage to Image/sourcedImage
The whole purpose of UnparsedImage is to not parse the manifest,
so having this interface which more or less requires the manifest to be
parsed and recomputed is not desirable.

sourcedImage now uses ImageSource.LayerInfosForCopy via the private
UnparsedImage.src, which is a bit ugly, but that dependency has already
existed, and we get a more consistent public API for this price.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>