Load a *storage.Image only once in storageImageSource
... instead of reloading it in LayerInfosForCopy and getSize.

Alternatively, it seems we could just save the TopLayer field; this
is more generic.

AFAICT this should be equally safe in principle: any locking done in
the storage.Store loaders is unlocked by the time we get the
*storage.Image, so if anything can modify the unlocked *storage.Image
concurrently after this change, or make its fields invalid, the same
breakage can in principle happen before this change just as well,
"only" with a shorter window.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>