Fix failure to make auth directory
Whenever the auth directory was defaulting to /run/usr/UID
it was assuming that /run/usr/UID existed. The change creates the
directory if it doesn't exist.
We were also creating the XDG_RUNTIME_DIR deirectory if the variable
was set but the directory didn't exist. Not a good idea so we error out
now when that happens.

Signed-off-by: umohnani8 <umohnani@redhat.com>