Remove all users of storageReference.name
They are mostly used only for boolean presence checking; only DockerReference
actually uses the value; in that case the transition to completeReference
may add a digest value, but a few lines below we do a, now redundant,
WithDigest computation which would have added the same digest anyway.

NOTE: This changes PolicyConfigurationIdentity and PolicyConfigurationNamespaces,
which now return values even if breakDockerReference.  That seems to be the right
thing to do anyway.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>