RFC: Simplify parsing of image IDs/digests
Instead of first extracting fields, and only later deciding based
on their existence and contents how to interpret them, combine the
field extraction and interpretation into one step.  We may possibly
check the trailing field twice if it is a digest, but overall it's simpler.

NOTE: Perhaps I am overlooking a reason the code was organized this way;
specific error reporting behavior perhaps?

Signed-off-by: Miloslav Trmač <mitr@redhat.com>