errors.Wrapf doesn't return error when err is nil
In this commit you can see how at the beginning of the function
`getManifestDescriptor` it was being checked for an error was different than
nil and returning in that case, however, few lines later that error (remember,
with nil value) was being wrapped. Since the original error was nil, the
wrapping was as well returning a nil.

A regression tests was added.

This commit should close the following issue in skopeo as per @mtrmac
suggestion: github.com/projectatomic/skopeo/issues/496

Signed-off-by: Álex González <agonzalezro@gmail.com>