dockerClient: skip TLS verification if configured in registries
Make dockerClient sensitive to the configured registries by using the
sysregistriesv2 package.  Unless specified in the SystemContext, the TLS
config's InsecureSkipVerify flag now adheres to the registries Insecure
field (if specified in the registries.conf).  Notice, that the default
value remains false (i.e., TLS verification is performed).

SystemContext.DockerInsecureSkipTLSVerify is usually set by users of
this library (e.g., Skopeo and Podman).  To be able to distinguish
between a default value and a user-specified value, introduce a
types.OptionalBool type that is now used for DockerInsecureSkipTLSVerify.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>