Add a special case for GzippedEmptyLayer to storageImage.GetBlob
This is not necessary yet, but we will discard throwaway layers
in PutBlob, which means we will need to re-create something equivalent
when reproducing the image.

Use the same GzippedEmptyLayer that we use for explicit conversions
to schema1, with the same rationale: this particular blob is likely
to exist on registries already.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>