Fix requested token actions when deleting images
https://github.com/docker/distribution/blob/master/docs/spec/auth/scope.md
does not document the action necessary for deleting images.

For docker/distribution, we need "pull" to be able to read the manifest,
and "delete" to delete it.  Until docker/distribution commit
ccb839e0e30c3b6992fb4084dfd6550d0ddd4d1a (Jan 3 2017), the action checked
for deleting manifests was "*".

For Quay.io, only "push" is sufficient for both; and although
https://github.com/docker/distribution/blob/master/docs/spec/auth/jwt.md
says that requesting permissions that the server does not grant should
not be an error, quay.io refuses to even parse a request which contains
a "delete" action and does not grant any token.

This should not affect OpenShift, where the token is an API
token and does not actually include any scopes or permissions;
(that is also probably why noone has noticed before).

Overall, "*" seems to be the only common action specification;
luckily both docker/distribution and Quay.io seem to treat it
as "all actions allowed" (which is _not_ documented in the spec,
but then again, nothing about actions for deleting images is
documented).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>