Update for recent changes to containers/storage
Modify storageImageDestination.Commit() to account for the recent change
to containers/storage which makes it so that the io.ReadCloser that
store.Diff() returns now holds a lock until it's closed.

Because that's the same lock that store.PutLayer() takes, we need to
finish reading the diff before we can do anything with its contents, and
we can't stream a layer's contents directly from one to an other.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>