Use a pool of progress bars
This is in preparation for copying layers concurrently, using a pool
will display a bar for every layer. An attempt has been made to align
the formatting of the bars with docker pull, within the constraints of
the cheggaaa/pb implementation, specifically:
- each bar is prefixed with the first 12 chars of the digest
- the bar updates with text for the current action (Copy, Skipping...)
- the bar style is now [=======>      ]

No pool is created for --debug, so each bar prints interleaved with
stderr, which resembles the current behaviour.

Even though the config is a blob, it doesn't get a progress bar (it
will be too small to really justify it)

So far this has been tested locally with skopeo copy.