sysregistriesv2: invalidate cache
Add the `InvalidateCache()` function to the sysregistriesv2 package to
allow users to explicitly invalidate the cache of registry config files.
Such functionality is required for long-running processes (e.g., CRI-O)
to reload potential changes made to the cached config files.

An alternative solution would be caching the modtime of the config files
at the cost of additional syscalls (plus eventually re-parsing), which
would introduce jitter that other container engines do not have.  In
contrast, the solution in this change makes reloading and hence the
costs of reloading explicit.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>