Add pkg/blobinfocache.DefaultCache()
This returns a default BoltDB instance (which almost everyone
should use to maximally benefit from the cache).  The default location
is chosen to exist near the containers/storage GraphRoot (i.e.
typically on the same partition, with the same SELinux context),
both for root and rootless modes of podman.

This adds no users yet, so should not immediately change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>