docker: mark localhost as insecure
Mark localhost as insecure to enable communicating with a registry
running on localhost without requiring TLS or explicitly marking it as
insecure in the registries.conf.  This behaviour aligns with Docker.

Notice, that localhost can still be marked to not skip TLS verification
either by setting it via the SystemContext or by configuring it
explicitly in the registries.conf.  A registry is considered to be
running on localhost if it either starts with "localhost:" or with
"127.0.0.1".

Fixes: #544
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>