tests: use corresponding skopeo repo for testing
Per default, use the master branch of containers/skopeo for testing.

But if there is a fork of skopeo in the same group as the
containers/image repository being tested and with the same branch,
then use it for testing. This enables testing containers/image PRs
modifying dependencies.

For example, for a request to pull branch `mybranch` of
`mygroup/image` (a fork of `containers/image`), a branch `mybranch` of
`mygroup/skopeo` (a fork of `containers/skopeo`) is needed. Of course,
if new dependencies are being added, an additional request will be
needed to pull branch `mybranch` of repository `mygroup/skopeo` to
branch `master` of repo `containers/skopeo`.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>