ostreeImageSource.GetBlob(): close mfz when we're done with it
The Reader that we use to read the decompressed copy of the tar split
data in ostreeImageSource.GetBlob() was being closed when we returned,
at which time the goroutine that was feeding the io.ReadCloser that we
returned could still be running.  Move its Close() call so that it's
closed along with the io.ReadCloser that we return.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>