Remove bar != nil checks
These checks were added to work around progressPool.AddBar()
returning nil after progressPool.Wait() was called, which could
happen in copyUpdatedConfigAndManifest (which called the .Wait())
was called more than once in a row.

We have fixed the lifetime of the progress pool, so this should no
longer be necessary.  (Also, it was very confusing, because
copyBlobFromStream was always calling bar.ProxyReader() -
so the code could crash on nil accesses anyway, unless PutBlob happened
to detect the blob existing at the destination and not read any data
from the proxy reader.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>