storage destination: blob-copy detection
Make use of the digest locks from c/storage to detect if other processes
are copying a given blob to effectively avoid redundantly copying the
same blob to the storage.

The heuristic to decide if a blob is currently being copied is to first
acquire the digest lock and then wait for 200 ms; if the lock hasn't
been acquired until then, we assume that another process is holding the
lock and is hence copying the blob.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>