Only log the result of reference rewriting if the result is valid
... which is not, strictly speaking, guaranteed if reference.ParseNamed
returns an error; and in that case we want to log the error and not really
the result, if any, anyway.  Also, don't bother checking whether the result
is not nil on success, it must be: the caller isn't ready to handle nil,
and there is no good way anyway; returning a real error value is the
resposibility of reference.ParseNamed.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>