Release 2.0
This is a major version bump because the API changes; while we will
try to minimize API changes in the future, past experience shows that
it is still extremely likely that they will be very frequent, unless
the API is significantly restructured and/or large parts of it hidden
or declared out of bounds.  Expect the major number to increase
several times before this can settle.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>