pkg/blobinfocache: fixup unit tests for euid 0
When testing, force our rootless ID to -1 to avoid using the hard-coded
default location when the tests are being run by a privileged user.
Turn off CAP_DAC_OVERRIDE in the test so that when we expect to be
denied the ability to create a subdirectory under a directory to which
DAC says we can't write, that's what will actually happen, even for
root.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>