zstd: support docker v2s2
Support zstd for docker v2s2 images.  That's required to copy and
convert a v2s2 image into an OCI images, for instance, via

   `skopeo copy --dest-compress --dest-compress-format=zstd docker://...  oci:...`

When compressing or decompressing a layer, we first update the current
manifest and later convert it.  Hence, supporting v2s2 is currently
required to support compressing and converting in one call.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>