Revert "update dependencies"
This reverts commit 3e4758b5bcaf6c22fbb512743a608f6ba271e680, which
was mostly _reverting_ dependencies.

(The reason for that commit was that mergo has incompatibly changed
behavior, causing atomic: cluster accesses to use localhost:8080
instead of the cluster API endpoint defined in the configuration.
The previous two commits have updated mergo and adjusted the callers,
so we no longer need to hold mergo back.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>