update github.com/vbauerster/mpb to v4
This is a purely janitorial change but paves the path to make use of all
improvements in v4 that we will need at some point, for instance, for
better replacing progress bars and updating their progress.

Use the decor.Name to replace on completion as the appended decorators
are now introducing whitespaces on completion; it also reduces the
number of decorators for spinners.

Fixes: #745
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>