Make restTLSClientConfig in restConfig a public field
imdario/mergo, despite its official documentation, does not merge
public fields inside private fields as of v.0.3.9:
https://github.com/imdario/mergo/issues/139

Fixing that seems non-trivial.

Instead, make the restTLSClientConfig a public field of restConfig.

The restConfig type itself remains private, so this does not make anything
actually public outside the subpackage. This way, the calls to mergo work
as expected with both 0.3.8 and 0.3.9.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>