Simplify the WantedPlatforms main loop
Instead of dealing with full imgspecv1.Platform, where the
arch/OS is always the same, just implement the logic in terms of
variants, and convert to imgspecv1.Platform separately.

This shortens the code, at the cost of introducing an extra array
which will be momentarily eliminated again in most cases.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>