allow ImageDestination to decide compression based on blob info
The various ImageDestinations know what types of layers are supported
natively and how they should be compressed, but some (e.g. the OCI
distribution repo or the OCI image spec) allow arbitrary blobs to be
uploaded. We should not compress these blobs by default, since we don't
know what they are and may already be compressed (e.g. squashfs).

So, let's introduce a new DesiredBlobCompression() and deprecate
DesiredLayerCompression() to allow the various backends to tell the core
code about what kind of compression they want.

We implement mime-type-based compression detection for the docker and OCI
backends. As above, OCI supports arbitrary blobs so we shouldn't interfere
with ones we don't understand. The docker backend is intended to be the way
we interact with OCI dist spec repos, so we should also not compress
unknown media types when talking to the docker backend.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>