Move docker/tarfile.Source to docker/internal/tarfile.Source
... so that we can add more functionality to it without making it
public API.

Keep the original docker/tarfile.Source as a compatibility shim.

ManifestItem is moved to the private package to avoid an import cycle,
but remains visible.

docker/tarfile/src_test.go should logically be moved as well, but we
can't do that yet due to the import cycle.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>