storage: check the compressed digest of cached blobs
When there are diffent images that have the same layer compressed
differently, pulling the images in cri-o may fail.
Reproducer:
 # crictl pull docker.io/fgiudici/test:1.0
 # crictl rmi docker.io/fgiudici/test:1.0
 # crictl pull docker.io/node:lts-slim
 # crictl pull docker.io/fgiudici/test:1.0
FATA[0003] pulling image failed: rpc error: code = Unknown desc = Error: blob sha256:6a73b40493d9a3fd12c93f0d03ec90c78e447dc9f182d119fb7477b75997ba72 is already present, but with size 22522274 instead of 23625492

Image docker.io/fgiudici/test:1.0 has some layers in common with docker.io/node:lts-slim, but the layers have been pushed with a different compressed archive (and so different compressed-diff-digest).
In particular:
node:lts-slim has
    "id": "767a7b7a8ec530ae9a7b1c4eeed530d54ca7ef9f224fe207be42dc74f565587a",
    "created": "2020-09-23T10:27:27.399649267Z",
    "compressed-diff-digest": "sha256:abb454610128b028301ee40af387d31111a1e699e4ea424fd53186ff77067402",
    "compressed-size": 22522274,
    "diff-digest": "sha256:767a7b7a8ec530ae9a7b1c4eeed530d54ca7ef9f224fe207be42dc74f565587a",
    "diff-size": 58489344,
    "compression": 2

test:1.0 has
    "id": "767a7b7a8ec530ae9a7b1c4eeed530d54ca7ef9f224fe207be42dc74f565587a",
    "created": "2020-09-23T09:50:52.502675155Z",
    "compressed-diff-digest": "sha256:6a73b40493d9a3fd12c93f0d03ec90c78e447dc9f182d119fb7477b75997ba72",
    "compressed-size": 23625492,
    "diff-digest": "sha256:767a7b7a8ec530ae9a7b1c4eeed530d54ca7ef9f224fe207be42dc74f565587a",
    "diff-size": 58489344,
    "compression": 2

This is due to an incomplete cleaning of the storage cache when the
images are deleted: the mappings of the compressed digest images are
not cleaned. So, if the same layer, but with a different compression, is
later pulled, the old undeleted reference will match, returning a layer
with a different compressed digest and compressed size.

Don't blindly trust the storage cache, check we got what we need.

Signed-off-by: Francesco Giudici <fgiudici@redhat.com>