docker tarfile: skip from gofmt check
We want to avoid changes as much as possible in this old and stable
branch. Let's just skip the gofmt check for the docker/tarfile/src.go,
which would fail on the alphabetical order of the imports.
Goal is to have "make validate" and CI not failing on the 1.11 branch
without changing the source code.

Signed-off-by: Francesco Giudici <fgiudici@redhat.com>