Sort Registries in V2RegistriesConf.postProcess
loadConfig sorts registries after loading any file, as
a side effect of its merging code; that code will in the future
not be called if there are no drop-ins and there is nothing to merge.

So, sort the Registries array here as well; the code was already iterating
through them to ensure uniqueness, so combining the sorting and this code
is one of the cheapest places where it can happen.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>