copy: early commit of storage blobs
When copying an image to a storage destination, make sure to commit
downloaded layers as early as possible.  Since layers are committed in
sequence, we need to preserve the index of each layer: layer N can only
be committed to the storage when all N-1 previous layers are already
present in the storage, in the given order.

Two new methods are added to the storage destination to pass down the
index to TryResusingBlob and PutBlob.  The idea is fairly simple:

 * PutBlobWithOptions and TryReusingBlobWithOptions can "queue in" work.

 * Exaclty one caller of them may enter a second stage which is
   committing the previously queued in blobs to the storage while
   preserving the order.  If one caller is already in stage two,
   others will return early freeing up goroutines workers to continue
   pulling more blobs if needed.

Note: since this optimization is limited to storage destinations, we do
not need to extend the entire public API with methods that only some
types implement.  Hence, add an internal `ImageDestinationWithOptions`
interface to the new `internal/types` package.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>