manifest: GuessMIMEType: support OCI artifacts
If a schema version 2 manifest doesn't include a top-level media type,
default to "guessing" it's an OCI v1 manifest.

Previously, we defaulted to a Docker v2s2 manifest which is causing
issues for OCI artifacts which may ship a custom config media type.
In that case, we guessed it's a Docker image which is wrong.

To harden the code a bit further, add an additional media-type check for
the config to not only check for an ordinary OCI config but also a
Docker one.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>