search: catalog: return full match as first item
When searching via v2/catalog make sure that a full match is returned as
the first item.  The change is motivated by a conversation in a BZ [1]
on the rather undefined nature/semantics of searching images.

While this change does not fix the BZ, it may be a small improvement
when using the v2/catalog endpoint.

[1] bugzilla.redhat.com/show_bug.cgi?id=1976283

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>