Stop caching encrypted layers in docker destination
In a case where a non-encrypted images is committed to a registry after
an encrypted image is committed, this library may try to reuse
encrypted layers for the non-encrypted images if the images have common
layers. When the non-encrypted image is committed, there is no
information to recognize that cached layers are encrypted. As a result,
the non-encrypted image has encrypted layers unexpectedly in the
registry. This causes an error when the non-encrypted image is pulled
from the registry.

To avoid reusing encrypted layers for another image, this change stops
caching a layer if it is encrypted.

Signed-off-by: Hironori Shiina <shiina.hironori@jp.fujitsu.com>