Only call blobInfo.Digest.Validate() once in PutBlob
Digest.Validate() is somewhat costly, and we don't want to
_silently_ ignore invalid values; so, call it only once,
and later use just Digest != "".

In theory this changes semantics, invalid values are now
used instead of ignored. In practice this almost can't happen,
because except when decrypting layers (and except for non-c/image/copy
callers), that digest must have already been used to
initialize a copy.digestingReader, which enforces validity
and algorithm availability.

Also, the blobinfo.Digest != hasher.Digest() failure means that
valid but unavailable algorithms end up failing - so the only
possible _new_ failures are truly invalid digest values, which
we are fine with rejecting.

This location, not at the beginning of the function, is a bit
strange for doing validation, but it allows
future commits to be simpler, and it will turn out fine eventually.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>