Switch hasher.Hash() and file in the PutBlob pipeline
ioutils.WriteCounter requires a Writer to pass thorugh, and
we want to make hasher.Hash() optional. So make sure WriteCounter
always has a destination, so that we can avoid an io.Discard.

Also split the TeeReader construction into two lines, to make
the next commit simpler.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>