Consistently compare docker.Digest with "" without Digest.String()
The "" value is not really a valid digest.Digest value anyway
(it fails Digest.Validate()), so comparing String() with "" is,
if anything, _less_ valid use of the API than comparing the raw
digest.Digest with "".

So, consistently use the shorter and easier-to-read version
throughout the codebase.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>