Fix GetAllCredentials
Before this set of commits, GetAllCredentials was incorrectly
passing namespaced values and URLs to GetCredentials, which was
silently broken.

Within this set of commits, GetCredentials started accepting
namespaced values, but refusing URL keys; so, update the code
to use the "keys" naming, and to normalize URL keys.

This is a behavior change.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>