Record locations of blobs discovered by PutBlob but not TryReusingBlob
In the rare case where TryReusingBlob (is not called, or) doesn't find
a blob in the destination but the following PutBlob does, do record its
location in the BlobInfoCache.

To do that, split the relevant part of TryReusingBlob into a separate function
instead of calling all of TryReusingBlob with restrictive parameters (and in
particular, NoCache, which is why we were not recording the location).

This should currently only matter in the rare race of TryReusingBlob vs.
registry changes; https://github.com/containers/image/pull/1393 proposes
adding another case that would be much easier to hit.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>