Update golang.org/x/crypto, and silence warnings about openpgp
This is a fallback code; the primary recommendation is to use the gpgme mechanism
implementation, which is out-of-process and more appropriate for handling long-term private key material
than any Go implementation.
For this verify-only fallback, we haven't reviewed any of the
existing alternatives to choose; so, for now, continue to
use this frozen deprecated implementation.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>