Inline tryReusingBlobLocked into tryReusingBlobAsPending
It's the only caller now, and the semantics of "Locked"
vs. "AsPending" is hard to describe; just have one larger
function that does all kinds of lookups.  If it turns out
to be too large, we can split it to individual lookups,
instead of having the SrcRef case treated as uniquely special.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>