Eliminate private.ImageDestinationPartial
Instead, move PutBlobPartial into private.ImageDestination,
and add a SupportsPutBlobPartial() call.

Sadly, Go doesn't have optional interface methods
like Objective C.

In a sense, the type check is more elegant, but doing it
this way it requires an explicit choice about what to do
for every single new transport, or new private optional feature,
and it looks a bit better on the call site (and it might be
slightly more efficient).

The need for two boilerplate methods at every private
transport is ugly, but it will also go away as we add
an internal/implement mixin.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>