toml decoding: report unknown keys
Despite the desire in #1284 to report unknown-key errors (e.g., for
typos), only log them on the debug level.  We have made rather bad
experiences with containers.conf which initially logged these errors
on the warning level which caused an immense noise for users.

The underlying problem is that not all tools are updated in sync,
especially for non-{CentOS,Fedora,RHEL} distributions.  Config files
changes for tools with an updated conf will hence cause noise for others
etc.

However, logging these error on the debug level will still improve user
experience and be especially helpful in analyzing users issues.

Fixes: #1284
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>