Move c/image/image to c/image/internal/image
... to allow us to add more APIs using a new mechanism.

Currently c/image/image exposes images as types.Image, which we can't change;
so, let's define a new API that will return a different type that we can modify
as necessary.

For now, make that new API internal, at least until we fix the way edits
are bundled with data-to-be-edited in types.BlobInfo.

c/image/image continues to exist, but currently only trivially forwards
all calls to the moved c/image/internal/image.

As a minimal proof of concept, none of the users of c/image/image in the repo
have been modified; they now refer to the compatibility forwarders.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>