Introduce sourceStream, use it for the first pipeline stage
The purpose of sourceStream is to make it a bit easier
to split the blob copy pipeline stages into separate functions;
they will update a *sourceStream, which we can pass a single parameter.

For now, build it and use it to initialize a digestingReader; then
update other code that should consume the current pipeline data.

NOTE: That is a judgement call, it is not always obvious what data
was intended to be used and what is a bug.

As a geneeral principle, use srcInfo (the original unmodified data)
in user-visible error reports, notably using the original digest.
Otherwise, lean towards using srcStream.info.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>