signing: add a build tag to exclude signing from the build
The build currently fails if CGO_ENABLED=0:

./mechanism_other.go:10:21: undefined: ContainersImageSignatureRequiresCgo

Add a new build tag containers_image_disable_signing which, instead of failing
to build, compiles in a stub which returns runtime errors when attempting to
sign container images.

If the user does not have cgo enabled they have two options:

 - tag containers_image_openpgp enables the insecure OpenPGP implementation.
 - tag containers_image_disable_signing returns runtime errors when signing.

The default is to fail the build w/ an error clarifying why.

Signed-off-by: Christian Stewart <christian@paral.in>