WIP: docker/docker_image_src: Support partial signature lookups
Sometimes we find some signatures, but hit an error before deciding
we've retrieved "all the signatures".  But "all the signatures" isn't
a well-defined thing.  Additional signatures may become available at
any time.  This commit adjusts the signature-fetching paths to allow
for "I had some trouble, so I'm not completely satisfied, but these
are the signatures I did find...", and maybe one of those will be
acceptable to the caller.

Signed-off-by: W. Trevor King <wking@tremily.us>