storage source: Don't store small blobs on disk in GetBlob()
To avoid locking the store for a large time we currently make a copy
of the blob into TemporaryDirectoryForBigFiles (i.e. /var/tmp) when
getting a stream to the blob. During a simple "podman run" this
happens 6 time.

However, all those calls are for the config blob, which we don't stream
anyway, so the temporary copy is unnecessary. This commit moves the
check for a the config blob outside of `getBlobAndLayerID()` allowing
us to return early in the config blob case and avoid the temporary file.

Signed-off-by: Alexander Larsson <alexl@redhat.com>