storage: Immediately unlink tmpfile
All we need in this case is an anonymous temporary file, so
unlink the file immediately after creating it.  This avoids
leaking space if the caller forgets to call `Close()` (and the
process exits).

This would have mitigated https://github.com/containers/skopeo/pull/1837
which was a missing `Close()` in the skopeo proxy code.

(It'd be better to use Linux `O_TMPFILE`, but this path is
 portable and avoids a new dependency)

Signed-off-by: Colin Walters <walters@verbum.org>