Refactor copy.Image to sign using []*signer.Signer
Consolidate the format specifics and options handling to a single
place, and let the rest deal with an undifferentiated list of Signers.

This might be user-visible: failures e.g. on invalid key file path are now
exposed immediately, instead of only after copying all the large layers.

This is a step towards allowing callers to provide their own Signer
objects.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>