Do not preallocate regex in init program
We are trying to speed up startup time of apps based on
containers/image regex takes some time on every start
when in init. Moving this to a sync.Once should not
effect any change on the system.

Only changing internal regexp, since changing external would be a
breaking change to the library.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>