Return private.ReusedBlob from TryReusingBlobWithOptions
This is a subset of types.BlobInfo the transports
actually should deal with. Also tighten the semantics a bit,
to be much more explicit about what the transports are responsible for.

This allows us to stop tracking MIME types in storage.addedLayerInfo,
so simplify that.

Then fix the generic code so that blob annotations are not discarded
on blob reuse.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>