copy.Options: add an InstancePlatforms field
Add an InstancePlatforms field to the copy.Options structure.  When
asked to copy only a specific subset of instances, combine the Instances
list with the instances which match the platforms given in the
InstancePlatforms list, returning an error if we're unable to find an
instance which matches any of the InstancePlatforms.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>