Add expected configs to conversion tests
The configs are actually where _most_ of the conversion happens, so maintain
tests for the contents.

Currently the configs are "tested" by requiring a specific size and digest,
which is actually good enough to prevent changes, but it does not make it
possible to review what changes were allowed when updating the tests.  So now
we will have both.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>