Workaround go-md2man warning about HTML comments
go-md2man 2.0.2 (not 1.0.10) does not recognize HTML comments,
triggering a warning.

Work around that, just to have a silent build.

This uses a "comment" syntax by defining a reference-style link
( https://stackoverflow.com/a/20885980 ).

Alternatively, go-md2man _does_ recognize comments inline inside
a paragraph, so preceding the HTML comment with a space would probably
work as well.

Either way is _weird_, and we can't add a comment to explain the comment mechanism...
so use the more visually distinct alternative to hopefully minimize regressions;
using a leading space would tempt future readers to just remove the space,
thinking it's a mistake.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>