Fix a deadlock in image creation
https://github.com/containers/image/pull/731 introduced
a WaitGroup which is waited but never signaled, and a nonsensical
error reporting mechanism (where errors are pushed but don't terminate
future operations).

Instead, rely on PipeWriter.CloseWithError() to propagate failures
from the goroutine.

And move all of the body of the goroutine into a separate function,
so that we can use defer() to ensure CloseWithError() is _always_
called.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>