Try all manifest formats if the destinations accepts all of them
We might want to trigger a conversion to OCI if the source is not
OCI, but the destination already contains a Zstd version of a layer.

We can do that for destinations that express a restricted list of
manifest formats, so it is unexpected that completely unrestricted
destinations can't trigger a conversion, and just fail (right now),
or perhaps don't fail but unnecessarily upload (in the future), in that case.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>