Ensure uncompressed layers from c/storage are recorded as uncompressed on upload
When copying an uncompressed layer without modification, arrange
types.BlobInfo.CompressionOperation to be set to types.Decompress,
so that we edit the manifest to use an uncompressed MIME type
(and not, e.g. a zstd:chunked MIME type which might not be representable
in the destination at all).

This is all still gated by the
> 	if srcInfosUpdated || layerDigestsDiffer(srcInfos, destInfos) {
condition in copyLayers, so we don't do these manifest updates
frivolously.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>