Move implementation of impl.CandidateMatchesTryReusingBlobOptions to internal/manifest
We will want to call this from the pkg/blobinfocache code, and for that we need
to make it dependent on neither internal/private nor internal/blobinfocache
(otherwise the interface definitions would create a dependency loop).

This only moves ~unchanged code, should not change behavior.

Eventually CandidateMatchesTryReusingBlobOptions will be removed entirely.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>