copy: avoid rewriting list manifest containing unknown fields
When copying a list manifest we currently create a new one and compare
its serialized version to the re-serialized version of the original
manifest. If they are identical we then use the original manifest to
keep the same digest.

But if the original manifest contains unhandled fields, like
`artifactType` or `subject` currently, the copy will always differ.
As a result we always write a new version without these fields.

We probably can't assume that any new field defined in the future can
always be copied as is into another manifest. So using the serialized
copy for comparison will always keep creating issues when new fields
are defined.

Instead explicitly check for the relevant fields, that is the media
type and list of instances. If these have not changed we can use the
original manifest without having to care about any additional fields
it might be using.

Signed-off-by: Alban Bedel <alban.bedel@aerq.com>