Eliminate a numLayers variable
The compiler is smart enough that referencing
a local variable is not any faster than len(srcInfos).

If anything, keeping the extra variable might add to register pressure
because the compiler might not understand that the srcInfos change due to
LayerInfosForCopy does not change the number of layers.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>