copy.Image: improve wrapping Close errors
When commit 849dd7014 was written, Go did not have an ability to wrap
multiple errors. Since Go 1.20 it's possible by either using multiple
%w, or via errors.Join (which discards nil errors).

In a sense, this is what the code was doing before commit 849dd7014,
except we don't add extra spaces or parentheses -- as joined error,
when converted to a string, is represented with newlines added between
elements.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>