directory: simplify newImageDestination
The current code is unnecessarily complicated:
 - it does unnecessary checks if a file/directory exists before opening it;
 - it uses three helper functions not used anywhere else;
 - directory contents is read twice.

Untangle all this, making the code simpler.

Also, move the logic of initializing a destination directory to a
separate function.

Now, since the logic of handing version file is now encapsulated in a
single function, remove versionPath method and const version, as well
as the corresponding test.

This also fixes the subtle issue of using ref.path for constructing the
version file path, but ref.resolvedPath for all other operations.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>