copy: Image: add/use safeClose
Consolidate the [duplicated] code that handles and wraps the errors from
Close to a function, and call it twice.

Keep the code which deliberately converts an error from Close to a
string (rather than wrapping it) as we don't want those errors to be
unwrappable (those are not "primary" errors).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>