.cirrus.yml: use skip, not only_if
Cirrus CI emits a bunch of warnings (shown as check annotations in GitHub UI), for example:

> .cirrus.yml#L75
> task "Test" depends on task "validate", but their only_if conditions are different

> .cirrus.yml#L96
> task "Skopeo Test" depends on task "validate", but their only_if conditions are different

> .cirrus.yml#L158
> task "Total Success" depends on task "validate", but their only_if conditions are different

According to Cirrus CI documentation [1], skip should be used to skip a
task (and mark it as "passed") if it is not needed, whereas only_if
marks the task as failed if the condition is not met.

Switch from using only_if to skip, and revert conditions.

Let's see if it works.

[1]: https://cirrus-ci.org/guide/writing-tasks/#conditional-task-execution
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>