Add OWNERS file
This is a proposal.  We do not have to turn the bot on at this point and
also feel free to sit on this PR until we have an opportunity to discuss
more broadly.

I am also happy to make edits now or later.  I basically looked at
contributions in the last 12-24 months.

Signed-off-by: Brent Baude <bbaude@redhat.com>