Pass trustedLayerIdentityData into createNewLayer
Two different locations in the function need the data, and the
caller must have it available; so always passing it in simplifies
the implementation and removes an impossible error path.

This might hypothetically make layer reuse a bit worse, if we
happened to learn something for trustedLayerIdentityData from
processing other layers of the same image, but reusing the same
layer twice within an image should be rare.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>