docs: registries.conf: mention Podman Machine
After a wild ride helping to configure a mirror for docker.io for Podman
on the Mac, I feel we need to improve the docs a bit.

When running against a Podman Machine, registries.conf is read
server-side only.  That means that mirrors need to be configured inside
the virtual Linux machine.  In addition to that, things can start
falling apart when adding a config to the HOME directory of the default
non-root (core) users of tha machine.

That means mirrors should be configured globally in
/etc/containers/registries.conf[.d].

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>