ignore ENOENT errors when parsing .crt files
As always listing files in a dir to then read them is racy as the file
might have been removed in the meantime. Thus we must ignore ENOENT
errors when the file is opened.

Now here the code already did not cause an hard error but it will cause
a spurious warning in such case. There is really no need to log that as
it can cause flakes for podman.

Now there is the case here for .cert and .key files where both files
must be present for a valid config. Ignoring ENOENT there seems wrong as
it would hide a common misconfiguration where only one of the files
exists. That mean the race can still cause a failure when these files
are removed from the dir.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>