oci/layout: add GetLocalBlobPath()
The podman artifact store needs a method to get the actual local path on
disk to the blobs. This is needed so we can then bind mount the files
into the container, of course the caller must ensure we treat the files
as read only. The digest based nature of the store will break if the
files are modified.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>