Migrate from github.com/xeipuuv/gojsonschema to github.com/santhosh-tekuri/jsonschema/v5
github.com/xeipuuv/gojsonschema has last had a commit in Oct 2020.

So, follow what github.com/opencontainers/image-spec did, and move to
the same library that project uses.

Note that github.com/santhosh-tekuri/jsonschema/v6 already exists,
but the goal is to decrease the dependency surface (to the little extent
this matters, it's at most a test dependency anyway), so we prioritize
using the older version.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>